pyproject.toml
[project]
name = "clinical-trials-dashboard"
version = "0.1.0"
description = "Clinical Trials Investment Dashboard using LangGraph and Streamlit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.38",
    "langgraph>=0.2",
    "langchain-openai>=0.2",
    "langchain-core>=0.3",
    "httpx>=0.27",
    "supabase>=2.0",
    "plotly>=5.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "exa-py>=1.0",
]
